--- crates/plugin_resolve/src/resolver_parser/mod.rs
@@ -1,1 +1,0 @@
-
