--- crates/plugin_resolve/src/resolver_parser/package_field.rs
@@ -1,4 +1,0 @@
-struct ParserPackageField<'a> {
-    pub name: &'a str,
-    pub value: &'a str,
-}
